[riscv][upstream] Make sifive_test finisher 4 bytes
QEMU is now stricter with MMIO sizes and accesses. size_t on RV64
is 8 bytes and generates an sd instruction that Store/AMO faults
because sifive_test MMIO expects 4 bytes accesses.

+++ bsps/riscv/riscv/start/bsp_fatal_halt.c
@@ -39,7 +39,7 @@ void _CPU_Fatal_halt(uint32_t source, uint32_t error)
 {
   const char *fdt;
   int node;
-  volatile size_t *sifive_test;
+  volatile uint32_t *sifive_test;
 
 #if RISCV_ENABLE_HTIF_SUPPORT != 0
   htif_poweroff();
